adds Transcript.biotype #244
